Remove TrailingCommaRuleTest completely
