+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/TrailingCommaOnCallSiteRule.kt
@@ -315,4 +315,3 @@ public class TrailingCommaOnCallSiteRule :
         )
     }
 }
-
